requirements.txt
dlib==19.23.0
Flask==1.1.2
Flask_Cors==3.0.10
Flask_Mail==0.9.1
Flask_SQLAlchemy==2.5.1
numpy==1.21.5
opencv_contrib_python==4.5.5.62
Pillow==9.2.0
requests==2.27.1
SQLAlchemy==1.4.31
timm==0.5.4
tlt==0.2.0
torch==1.11.0
